Color is now coded on one byte. That does not save so much memory... Bug in the shade computation.

+++ src/drawable.h
@@ -27,11 +27,11 @@ union Color
 {
     struct
     {
-        float r;
-        float g;
-        float b;
+        GLubyte r;
+        GLubyte g;
+        GLubyte b;
     };
-    float v[3];
+    GLubyte v[3];
 };
 
 class Drawable : public QObject
